setup angular-in-memory-web-api and learn about it
