+++ package.json
@@ -3,11 +3,12 @@
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve",
+    "start": "ng serve --proxy-config proxy.conf.json",
     "build": "ng build",
     "test": "ng test",
     "lint": "ng lint",
-    "e2e": "ng e2e"
+    "e2e": "ng e2e",
+    "server": "node node_modules/ngf-server/server.js"
   },
   "private": true,
   "dependencies": {
@@ -21,9 +22,11 @@
     "@angular/router": "~11.2.11",
     "@ngx-translate/core": "^13.0.0",
     "@ngx-translate/http-loader": "^6.0.0",
+    "angular-in-memory-web-api": "^0.11.0",
     "bootstrap": "^5.0.1",
     "bootstrap-icons": "^1.5.0",
     "jquery": "^3.6.0",
+    "ngf-server": "^1.0.3",
     "rxjs": "~6.6.0",
     "tslib": "^2.0.0",
     "zone.js": "~0.11.3"
@@ -35,7 +38,6 @@
     "@types/jasmine": "~3.6.0",
     "@types/jquery": "^3.5.5",
     "@types/node": "^12.11.1",
-    "angular-in-memory-web-api": "^0.11.0",
     "codelyzer": "^6.0.0",
     "jasmine-core": "~3.6.0",
     "jasmine-spec-reporter": "~5.0.0",
