+++ proxy.conf.json
@@ -1,0 +1,7 @@
+{
+  "/api/*": {
+    "target": "http://localhost:8808",
+    "secure": false,
+    "logLevel": "debug"
+  }
+}
