+++ src/app/app-routing.module.ts
@@ -7,19 +7,24 @@ import { UserDetailResolver } from './user/resolver/user-detail/user-detail.reso
 const routes: Routes = [
   {
     path: 'user', component: UserListComponent, resolve: {
-      data: UserListResolver
+      developers: UserListResolver
+    }
+  },
+  {
+    path: 'user/new/:id', component: UserCreateComponent, canActivate: [RouteActivedGuard], canDeactivate: ['canDeactivateCreateUser'], resolve: {
+      developer: UserDetailResolver
     }
   },
   { path: 'user/new', component: UserCreateComponent, canDeactivate: ['canDeactivateCreateUser'] },
   {
     path: 'user/:id', component: UserDetailComponent, canActivate: [RouteActivedGuard], resolve: {
-      data: UserDetailResolver
+      developer: UserDetailResolver
     }
   },
   { path: '404', component: Component404Component },
   { path: 'profile', loadChildren: () => import('./profile/profile.module').then(m => m.ProfileModule) },
   { path: '', redirectTo: '/user', pathMatch: 'full' },
-  { path: '**', redirectTo: '404',  },
+  { path: '**', redirectTo: '404', },
 ];
 
 @NgModule({
