+++ src/app/shared/directives/trigger-modal/trigger-modal.directive.ts
@@ -1,4 +1,5 @@
 import { Directive, ElementRef, Inject, Input, OnInit } from '@angular/core';
+import { timeout } from 'rxjs/operators';
 import { JQUERY_TOKEN } from 'src/app/components';
 
 @Directive({
@@ -15,8 +16,9 @@ export class TriggerModalDirective implements OnInit {
 
   ngOnInit() {
     this.el.addEventListener('click', (e) => {
-
-      this.$(`#${this.modalId}`).modal('show')
+      setTimeout(() => {
+        this.$(`#${this.modalId}`).modal('show');
+      }, 200);
     })
   }
 }
