+++ src/app/shared/index.ts
@@ -1,0 +1,7 @@
+export * from './directives/trigger-modal/trigger-modal.directive';
+export * from './directives/lazy-img/lazy-img.directive';
+export * from './directives/validator/validator.directive';
+export * from './pipes/read-enum/read-enum.pipe';
+export * from './service/data/data.service';
+export * from './simpleModal.component'
+export * from './shared.module'
