+++ src/app/user/resolver/user-list/user-list.resolver.ts
@@ -1,11 +1,14 @@
 import { Injectable } from '@angular/core';
 import {
-  Router, Resolve,
+  Resolve,
   RouterStateSnapshot,
   ActivatedRouteSnapshot
 } from '@angular/router';
+import { Subject } from 'rxjs';
+import { ObservableInput } from 'rxjs';
 import { Observable } from 'rxjs';
-import { map } from 'rxjs/operators';
+import { catchError, finalize, map } from 'rxjs/operators';
+import { IUser } from '../../models';
 import { UserService } from '../../service/user/user.service';
 
 @Injectable()
@@ -14,8 +17,11 @@ export class UserListResolver implements Resolve<boolean> {
 
   }
 
+  // catchError((err, d: Observable<IUser[]>): ObservableInput<any> => { return new Subject() })
+
   resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<any> {
-    return this._userService.getUSer().pipe(map(data => data));
+    return this._userService.getUSer()
+      .pipe(map(data => data));
   }
 
 }
