+++ src/app/user/user-create/user-create.component.html
@@ -3,7 +3,7 @@
 </h3>
 <hr>
 <div class="col-md-6 col-sm-12">
-  <form [formGroup]="formDeveloper" (ngSubmit)="createUser(formDeveloper.value)">
+  <form [formGroup]="formDeveloper" (ngSubmit)="save(formDeveloper.value)">
 
     <fieldset>
       <legend>Personnal information</legend>
