+++ src/styles.scss
@@ -16,7 +16,7 @@ body {
 }
 
 .bi {
-  font-size: 2rem;
+  // font-size: 1rem;
   &.middle {
     color: rgba(243, 44, 44, 0.897) !important;
   }
